TESTING: there seem to be rare edge cases (appears randomly) where postgres tests fail because of some database setup issues.
Refactor test_monitor_handler.py: Refactor reset_db() (db_cleanup)
function. In some cases there are issues, when running tests where
the public schema cannot be found anymore or seems not the be set
as default schema. THIS PROBLEM STILL OCCURES SOMETIMES. I was not able
to reproduce it / find the cause appropriately. I hope it is a local
db error but in case there are bugs concerning this in the future,
this function could be a place to start at.

+++ tests/test_monitor_handler.py
@@ -14,19 +14,22 @@ from pytest_monitor.sys_utils import determine_scm_revision
 from pytest_monitor.handler import SqliteDBHandler
 from pytest_monitor.handler import PostgresDBHandler
 
+import sys
+
 # helper function
 def reset_db(cleanup_cursor):
-    try:
+    cleanup_cursor.execute("SELECT schema_name FROM information_schema.schemata WHERE schema_name = 'public';")
+    has_schema = cleanup_cursor.fetchone()
+
+    if has_schema is not None:
         cleanup_cursor.execute("DROP SCHEMA public CASCADE;")
-        cleanup_cursor.execute("CREATE SCHEMA public;")
-        cleanup_cursor.execute("SET search_path TO public;")
-        cleanup_cursor.execute("SET schema public;")
-        cleanup_cursor.execute("ALTER ROLE postgres SET search_path TO public;")
-        cleanup_cursor.execute("ALTER SCHEMA public OWNER to postgres;")
-        cleanup_cursor.execute("GRANT ALL ON SCHEMA public TO postgres;")
-        cleanup_cursor.execute("GRANT ALL ON SCHEMA public TO public;")
-    except Exception:
-        pass
+
+    cleanup_cursor.execute("CREATE SCHEMA public;")
+    cleanup_cursor.execute("SET search_path TO public;")
+    cleanup_cursor.execute("ALTER ROLE postgres SET search_path TO public;")
+    cleanup_cursor.execute("ALTER SCHEMA public OWNER to postgres;")
+    cleanup_cursor.execute("GRANT ALL ON SCHEMA public TO postgres;")
+    cleanup_cursor.execute("GRANT ALL ON SCHEMA public TO public;")
 
 @pytest.fixture
 def sqlite_empty_mock_db() -> sqlite3.Connection:
